+++ src/DevilDaggersInfo.Core.Common/Extensions/TimeExtensions.cs
@@ -1,15 +1,35 @@
 namespace DevilDaggersInfo.Core.Common.Extensions;
 
-// TODO: Refactor to fixed-time struct.
+[Obsolete("Use GameTime struct instead.")]
 public static class TimeExtensions
 {
-	public static int To10thMilliTime(this float time) => (int)(time * 10000.0);
+	[Obsolete("Use GameTime struct instead.")]
+	public static int To10thMilliTime(this float time)
+	{
+		return (int)(time * 10000.0);
+	}
 
-	public static int To10thMilliTime(this double time) => (int)(time * 10000.0);
+	[Obsolete("Use GameTime struct instead.")]
+	public static int To10thMilliTime(this double time)
+	{
+		return (int)(time * 10000.0);
+	}
 
-	public static double ToSecondsTime(this int time) => time * 0.0001;
+	[Obsolete("Use GameTime struct instead.")]
+	public static double ToSecondsTime(this int time)
+	{
+		return time * 0.0001;
+	}
 
-	public static double ToSecondsTime(this uint time) => time * 0.0001;
+	[Obsolete("Use GameTime struct instead.")]
+	public static double ToSecondsTime(this uint time)
+	{
+		return time * 0.0001;
+	}
 
-	public static double ToSecondsTime(this ulong time) => time * 0.0001;
+	[Obsolete("Use GameTime struct instead.")]
+	public static double ToSecondsTime(this ulong time)
+	{
+		return time * 0.0001;
+	}
 }