+++ src/DevilDaggersInfo.Core.CriteriaExpression/Parts/ExpressionValue.cs
@@ -1,5 +1,4 @@
 using DevilDaggersInfo.Core.Common;
-using DevilDaggersInfo.Core.Common.Extensions;
 using DevilDaggersInfo.Core.Wiki;
 
 namespace DevilDaggersInfo.Core.CriteriaExpression.Parts;
@@ -17,6 +16,6 @@ public record ExpressionValue(int Value) : IExpressionPart
 			return Deaths.GetDeathByType(GameConstants.CurrentVersion, (byte)Value)?.Name ?? "???";
 
 		bool isTime = criteriaType is CustomLeaderboardCriteriaType.Time or CustomLeaderboardCriteriaType.LevelUpTime2 or CustomLeaderboardCriteriaType.LevelUpTime3 or CustomLeaderboardCriteriaType.LevelUpTime4;
-		return isTime ? Value.ToSecondsTime().ToString(StringFormats.TimeFormat) : Value.ToString();
+		return isTime ? GameTime.FromGameUnits(Value).Seconds.ToString(StringFormats.TimeFormat) : Value.ToString();
 	}
 }