+++ src/test/DevilDaggersInfo.Core.Common.Test/TimeTests.cs
@@ -1,5 +1,3 @@
-using DevilDaggersInfo.Core.Common.Extensions;
-
 namespace DevilDaggersInfo.Core.Common.Test;
 
 [TestClass]
@@ -8,8 +6,10 @@ public class TimeTests
 	[TestMethod]
 	public void TestTime()
 	{
-		Assert.AreEqual(10001998, 1000.1998.To10thMilliTime());
-		Assert.AreEqual(1.0, 10000.ToSecondsTime());
-		Assert.AreEqual(1234567890.1234, 12345678901234UL.ToSecondsTime());
+		Assert.AreEqual(1000.1998, GameTime.FromGameUnits(10001998).Seconds);
+		Assert.AreEqual(1.0, GameTime.FromSeconds(1.0).Seconds);
+		Assert.AreEqual(1.0, GameTime.FromSeconds(1).Seconds);
+		Assert.AreEqual(1234567890.1234, GameTime.FromGameUnits(12345678901234UL).Seconds);
+		Assert.AreEqual(1234567890.1234, GameTime.FromGameUnits(12345678901234L).Seconds);
 	}
 }